rearranged src folder structure
