+++ src/core/App.jsx
@@ -1,3 +1,8 @@
+/**
+ * App.js
+ * The top-level view of the application.
+ */
+
 import React from 'react';
 import {Navbar} from './components/Navbar.jsx';
 import {Footer} from './components/Footer.jsx';
