+++ src/core/components/AuthPage.jsx
@@ -1,5 +1,10 @@
+/**
+ * AuthPage.jsx
+ * Wrapper component that shows either it's children, or the login page if the user is not logged in.
+ */
+
 import React from 'react';
-import { LoginForm } from '../components/LoginForm.jsx';
+import { LoginForm } from './LoginForm.jsx';
 import { AppStateContext } from '../App.jsx';
 
 export function AuthPage({children}){
