+++ src/core/components/Breadcrumbs.jsx
@@ -1,3 +1,7 @@
+/**
+ * BreadCrumbs.jsx
+ * A breadcrumbs component.
+ */
 
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
