+++ src/core/components/Footer.jsx
@@ -1,3 +1,8 @@
+/**
+ * Footer.jsx
+ * Shows a simple footer component with links.
+ */
+
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCode } from '@fortawesome/free-solid-svg-icons';
