+++ src/core/views/Home.jsx
@@ -1,11 +1,17 @@
+/**
+ * Home.jsx
+ * The initial landing page
+ */
 
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCode } from '@fortawesome/free-solid-svg-icons';
+import { PostsSummary } from '../components/PostsSummary.jsx';
 
-export function NotFound(){
+export function Home(){
 	return (<div>
 		<h1><FontAwesomeIcon icon={faCode} /> Fuckit</h1>
-		<p>404: Page not found.</p>
+		<p>A simple PHP, SQLite, and ReactJS CMS for programmers.</p>
+		<PostsSummary />
 	</div>);
 }
