+++ src/core/views/NotFound.jsx
@@ -1,12 +1,15 @@
+/**
+ * NotFound.jsx
+ * The 404/page not found page. 
+ */
+
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCode } from '@fortawesome/free-solid-svg-icons';
-import { PostsSummary } from '../components/PostsSummary.jsx';
 
-export function Home(){
+export function NotFound(){
 	return (<div>
 		<h1><FontAwesomeIcon icon={faCode} /> Fuckit</h1>
-		<p>A simple PHP, SQLite, and ReactJS CMS for programmers.</p>
-		<PostsSummary />
+		<p>404: Page not found.</p>
 	</div>);
-} 
+}
