+++ src/core/views/Post.jsx
@@ -1,3 +1,8 @@
+/**
+ * Post.jsx
+ * The page that displays a post.
+ */
+
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCode } from '@fortawesome/free-solid-svg-icons';
